--- apps/web/src/entities/auction/queries/index.md
@@ -1,1 +1,0 @@
-// tanstackQuery
